+++ examples/v2/downloadBlob.ts
@@ -11,22 +11,23 @@
  */
 
 import { mainline } from "../mainline.ts";
-import { parseRepoAndRef } from "../../lib/common.ts";
-import { RegistryClientV2 } from "../../lib/registry-client-v2.ts";
+import { parseRepoAndRef } from "@cloudydeno/docker-registry-client/common";
+import { RegistryClientV2 } from "@cloudydeno/docker-registry-client/registry-client-v2";
 
 // Shared mainline with examples/foo.js to get CLI opts.
-const {opts, args} = mainline({ cmd: 'downloadBlob' });
+const cmd = 'downloadBlob';
+const {opts, args} = mainline({ cmd });
 if (!args[0] || (args[0].indexOf(':') === -1 && !args[1])) {
-    console.error('usage: ./examples/v2/%s.ts REPO@DIGEST');
+    console.error('usage: examples/v2/%s.ts REPO@DIGEST', cmd);
     Deno.exit(2);
 }
 
 // The interesting stuff starts here.
-var rar = parseRepoAndRef(args[0]);
+const rar = parseRepoAndRef(args[0]);
 if (!rar.digest) throw new Error('must specify a @DIGEST');
 console.log('Repo:', rar.canonicalName);
 
-var client = new RegistryClientV2({
+const client = new RegistryClientV2({
     repo: rar,
     insecure: opts.insecure,
     username: opts.username,
@@ -35,7 +36,7 @@ var client = new RegistryClientV2({
 
 const {ress, stream} = await client.createBlobReadStream({digest: rar.digest});
 
-var filename = rar.digest.split(':')[1].slice(0, 12) + '.blob';
+const filename = rar.digest.split(':')[1].slice(0, 12) + '.blob';
 console.log('Downloading blob to "%s".', filename);
 console.log('Response headers:');
 console.table(Array.from(ress[0].headers));
@@ -44,10 +45,6 @@ if (ress.length > 1) {
     console.table(Array.from(ress[ress.length - 1].headers));
 }
 
-const file = await Deno.create(new URL(filename, import.meta.url));
-for await (const buf of stream) {
-    await Deno.writeAll(file, buf);
-}
-file.close();
+await Deno.writeFile(new URL(filename, import.meta.url), stream);
 
 console.log('Done downloading', filename);