+++ examples/v2/getManifest.ts
@@ -11,12 +11,13 @@
  */
 
 import { mainline } from "../mainline.ts";
-import { parseRepoAndRef } from "../../lib/common.ts";
-import { RegistryClientV2 } from "../../lib/registry-client-v2.ts";
+import { parseRepoAndRef } from "@cloudydeno/docker-registry-client/common";
+import { RegistryClientV2 } from "@cloudydeno/docker-registry-client/registry-client-v2";
 
 // Shared mainline with examples/foo.js to get CLI opts.
+const cmd = 'getManifest';
 const {opts, args} = mainline({
-    cmd: 'getManifest',
+    cmd,
     options: [{
         names: ['accept-list'],
         type: 'bool',
@@ -29,7 +30,7 @@ const {opts, args} = mainline({
 });
 const name = args[0];
 if (!name) {
-    console.error('usage: node examples/v2/%s.js REPO[:TAG|@DIGEST]\n');
+    console.error('usage: examples/v2/%s.ts REPO[:TAG|@DIGEST]\n', cmd);
     Deno.exit(2);
 }