+++ examples/v2/headBlob.ts
@@ -11,14 +11,15 @@
  */
 
 import { mainline } from "../mainline.ts";
-import { parseRepoAndRef } from "../../lib/common.ts";
-import { RegistryClientV2 } from "../../lib/registry-client-v2.ts";
+import { parseRepoAndRef } from "@cloudydeno/docker-registry-client/common";
+import { RegistryClientV2 } from "@cloudydeno/docker-registry-client/registry-client-v2";
 
 // Shared mainline with examples/foo.js to get CLI opts.
-const {opts, args} = mainline({cmd: 'headBlob'});
+const cmd = 'headBlob';
+const {opts, args} = mainline({ cmd });
 const name = args[0];
 if (!name) {
-    console.error('usage: node examples/v2/%s.js REPO@DIGEST');
+    console.error('usage: examples/v2/%s.ts REPO@DIGEST', cmd);
     Deno.exit(2);
 }