+++ test/common.test.ts
@@ -8,9 +8,10 @@
  * Copyright (c) 2015, Joyent, Inc.
  */
 
-import { assertEquals, assertThrows, assertObjectMatch } from "https://deno.land/std@0.130.0/testing/asserts.ts";
-import { parseIndex, parseRepo, parseRepoAndRef, parseRepoAndTag } from "../lib/common.ts";
-import { RegistryIndex } from "../lib/types.ts";
+import { assertEquals, assertThrows, assertObjectMatch } from "./util.ts";
+
+import { parseIndex, parseRepo, parseRepoAndRef, parseRepoAndTag } from "@cloudydeno/docker-registry-client/common";
+import { RegistryIndex } from "@cloudydeno/docker-registry-client/types";
 
 // --- Tests