+++ test/v2.other.test.ts
@@ -4,8 +4,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { assertEquals } from "https://deno.land/std@0.130.0/testing/asserts.ts";
-import { digestFromManifestStr } from "../lib/registry-client-v2.ts";
+import { assertEquals } from "./util.ts";
+import { digestFromManifestStr } from "@cloudydeno/docker-registry-client/registry-client-v2";
 
 /*
  * Copyright (c) 2016, Joyent, Inc.