+++ test/v2.quayio.test.ts
@@ -15,9 +15,9 @@ import {
     dirname,
 } from "./util.ts";
 
-import { RegistryClientV2 } from "../lib/registry-client-v2.ts";
-import { parseRepo, MEDIATYPE_OCI_MANIFEST_V1 } from "../lib/common.ts";
-import { ManifestOCI } from "../lib/types.ts";
+import { RegistryClientV2 } from "@cloudydeno/docker-registry-client/registry-client-v2";
+import { parseRepo, MEDIATYPE_OCI_MANIFEST_V1 } from "@cloudydeno/docker-registry-client/common";
+import { ManifestOCI } from "@cloudydeno/docker-registry-client/types";
 
 // --- globals